+++ R/webservice_functions.R
@@ -127,11 +127,11 @@ token_ws <- function(username = NULL, password = NULL) {
 
 realtime_ws <- function(station_number, parameters = c(46, 16, 52, 47, 8, 5, 41, 18),
                         start_date = Sys.Date() - 30, end_date = Sys.Date(), token) {
-  if (length(station_number) >= 300) {
-    stop("Only 300 stations are supported for one request. If more stations are required,
-         a separate request should be issued to include the excess stations. This second request can
-         be issued on the same token if it isn't required.")
-  }
+  #if (length(station_number) >= 300) {
+  #  stop("Only 300 stations are supported for one request. If more stations are required,
+  #       a separate request should be issued to include the excess stations. This second request can
+  #       be issued on the same token if it isn't required.")
+  #}
 
 
   ## Check to see if the token is expired
@@ -214,10 +214,11 @@ realtime_ws <- function(station_number, parameters = c(46, 16, 52, 47, 8, 5, 41,
 
   csv_df <- dplyr::left_join(
     csv_df,
-    dplyr::select(param_id, -Name_Fr),
+    dplyr::select(tidyhydat.ws::param_id, -.data$Name_Fr),
     by = c("Parameter")
   )
-  csv_df <- dplyr::select(csv_df, STATION_NUMBER, Date, Name_En, Value, Unit, Grade, Symbol, Approval, Parameter, Code)
+  csv_df <- dplyr::select(csv_df, .data$STATION_NUMBER, .data$Date, .data$Name_En, .data$Value, .data$Unit,
+                          .data$Grade, .data$Symbol, .data$Approval, .data$Parameter, .data$Code)
 
   ## What stations were missed?
   differ <- setdiff(unique(station_number), unique(csv_df$STATION_NUMBER))